My first attempt at a Crystal wrapper for GDAL
This is mostly the output of crystal_lib for the GDAL, OGR, and CPL
header files.  But I did add a few helper functions to map errors into
exceptions.

Still needs work as I learn the details of the require & name space
implementation in crystal.  I still think mostly in python-terms but
this work different here.