Move the Libs under the respective modules
In crystal (like ruby) everything is global by default.  So defining a
lib at the global level makes it visible to every bit of code in the
program--whether the include any of the relevant files or not.

To try to keep things clean, I moved the libs under the respective
modules.  So LibCPL is now CPL::Lib.